Unify stream ID types and event APIs across HTTP/1.1, HTTP/2, HTTP/3
Change H2 and HTTP/1.1 stream IDs from u32 to u64 to match the H3/QUIC
API surface, making protocol-agnostic code possible without casts. Add
Http1Event::Connected for parity with H2Event::Connected and H3Event.

+++ examples/h2_client.rs
@@ -31,7 +31,7 @@ fn main() {
     // Drive the handshake first.
     let mut handshake_done = false;
     let mut request_sent = false;
-    let mut response_stream: Option<u32> = None;
+    let mut response_stream: Option<u64> = None;
 
     for _round in 0..100 {
         // Send pending output.
