+++ src/h2/client.rs
@@ -46,7 +46,7 @@ impl<const MAX_STREAMS: usize, const BUF: usize, const HDRBUF: usize, const DATA
         authority: &str,
         extra_headers: &[(&[u8], &[u8])],
         end_stream: bool,
-    ) -> Result<u32, Error> {
+    ) -> Result<u64, Error> {
         let mut headers: heapless::Vec<(&[u8], &[u8]), 20> = heapless::Vec::new();
         let _ = headers.push((b":method", method.as_bytes()));
         let _ = headers.push((b":path", path.as_bytes()));
@@ -61,7 +61,7 @@ impl<const MAX_STREAMS: usize, const BUF: usize, const HDRBUF: usize, const DATA
     /// Send body data on a stream.
     pub fn send_body(
         &mut self,
-        stream_id: u32,
+        stream_id: u64,
         data: &[u8],
         end_stream: bool,
     ) -> Result<usize, Error> {
@@ -71,7 +71,7 @@ impl<const MAX_STREAMS: usize, const BUF: usize, const HDRBUF: usize, const DATA
     /// Read response headers.
     pub fn recv_headers<F: FnMut(&[u8], &[u8])>(
         &mut self,
-        stream_id: u32,
+        stream_id: u64,
         emit: F,
     ) -> Result<(), Error> {
         self.inner.recv_headers(stream_id, emit)
@@ -80,7 +80,7 @@ impl<const MAX_STREAMS: usize, const BUF: usize, const HDRBUF: usize, const DATA
     /// Read response body.
     pub fn recv_body(
         &mut self,
-        stream_id: u32,
+        stream_id: u64,
         buf: &mut [u8],
     ) -> Result<(usize, bool), Error> {
         self.inner.recv_body(stream_id, buf)
@@ -157,7 +157,7 @@ mod tests {
 
         // Server sees headers
         let mut got_headers = false;
-        let mut header_stream = 0u32;
+        let mut header_stream = 0u64;
         while let Some(ev) = server.poll_event() {
             if let H2Event::Headers(sid) = ev {
                 got_headers = true;
